# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LIMAPAY API - Gestion commerciale pour commerces physiques du Sénégal
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  owner: 'owner',
  cashier: 'cashier',
} as const;

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  shopName: string;
  role?: RegisterInputRole;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  owner: 'owner',
  cashier: 'cashier',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  shopName: string;
  createdAt?: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Product {
  id: number;
  name: string;
  salePrice: number;
  purchasePrice: number;
  stock: number;
  lowStockThreshold?: number;
  category: string;
  /** @nullable */
  barcode?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  createdAt?: string;
}

export interface ProductInput {
  name: string;
  salePrice: number;
  purchasePrice: number;
  stock: number;
  lowStockThreshold?: number;
  category: string;
  barcode?: string;
  photoUrl?: string;
}

export interface ProductUpdate {
  name?: string;
  salePrice?: number;
  purchasePrice?: number;
  stock?: number;
  lowStockThreshold?: number;
  category?: string;
  barcode?: string;
  photoUrl?: string;
}

export interface SaleItem {
  id?: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
}

export type SalePaymentMethod = typeof SalePaymentMethod[keyof typeof SalePaymentMethod];


export const SalePaymentMethod = {
  cash: 'cash',
  wave: 'wave',
  orange_money: 'orange_money',
  free_money: 'free_money',
  diamondpay: 'diamondpay',
  credit: 'credit',
} as const;

export type SaleStatus = typeof SaleStatus[keyof typeof SaleStatus];


export const SaleStatus = {
  completed: 'completed',
  pending: 'pending',
  cancelled: 'cancelled',
} as const;

export interface Sale {
  id: number;
  totalAmount: number;
  amountPaid?: number;
  changeAmount?: number;
  paymentMethod: SalePaymentMethod;
  status: SaleStatus;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  diamondPayTransactionId?: string | null;
  items: SaleItem[];
  createdAt?: string;
}

export interface SaleItemInput {
  productId: number;
  quantity: number;
}

export type SaleInputPaymentMethod = typeof SaleInputPaymentMethod[keyof typeof SaleInputPaymentMethod];


export const SaleInputPaymentMethod = {
  cash: 'cash',
  wave: 'wave',
  orange_money: 'orange_money',
  free_money: 'free_money',
  diamondpay: 'diamondpay',
  credit: 'credit',
} as const;

export interface SaleInput {
  items: SaleItemInput[];
  paymentMethod: SaleInputPaymentMethod;
  totalAmount: number;
  amountPaid?: number;
  customerId?: number;
  diamondPayTransactionId?: string;
}

export interface Customer {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  totalDebt?: number;
  createdAt?: string;
}

export interface CustomerInput {
  name: string;
  phone: string;
  email?: string;
}

export interface CustomerUpdate {
  name?: string;
  phone?: string;
  email?: string;
}

export interface DebtPayment {
  id: number;
  debtId: number;
  amount: number;
  /** @nullable */
  note?: string | null;
  paidAt: string;
}

export type DebtStatus = typeof DebtStatus[keyof typeof DebtStatus];


export const DebtStatus = {
  pending: 'pending',
  partial: 'partial',
  paid: 'paid',
} as const;

export interface Debt {
  id: number;
  customerId: number;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  saleId?: number | null;
  originalAmount: number;
  remainingAmount: number;
  status: DebtStatus;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  note?: string | null;
  payments?: DebtPayment[];
  createdAt: string;
}

export interface DebtInput {
  customerId: number;
  saleId?: number;
  originalAmount: number;
  dueDate?: string;
  note?: string;
}

export type DebtUpdateStatus = typeof DebtUpdateStatus[keyof typeof DebtUpdateStatus];


export const DebtUpdateStatus = {
  pending: 'pending',
  partial: 'partial',
  paid: 'paid',
} as const;

export interface DebtUpdate {
  dueDate?: string;
  note?: string;
  status?: DebtUpdateStatus;
}

export interface DebtPaymentInput {
  amount: number;
  note?: string;
}

export interface DashboardSummary {
  revenueToday: number;
  revenueWeek: number;
  revenueMonth: number;
  salesToday: number;
  salesWeek: number;
  salesMonth: number;
  averageBasket: number;
  totalDebtAmount: number;
  totalProducts: number;
  lowStockCount: number;
  outOfStockCount: number;
}

export interface TopProduct {
  productId: number;
  productName: string;
  totalQuantity: number;
  totalRevenue: number;
}

export interface SalesPeriodPoint {
  label: string;
  revenue: number;
  salesCount: number;
}

export interface StockAlerts {
  lowStock: Product[];
  outOfStock: Product[];
}

export type DiamondPayInitInputOperator = typeof DiamondPayInitInputOperator[keyof typeof DiamondPayInitInputOperator];


export const DiamondPayInitInputOperator = {
  wave: 'wave',
  orange_money: 'orange_money',
  free_money: 'free_money',
} as const;

export interface DiamondPayInitInput {
  amount: number;
  saleId?: number;
  operator?: DiamondPayInitInputOperator;
  phoneNumber?: string;
}

export type DiamondPayTransactionStatus = typeof DiamondPayTransactionStatus[keyof typeof DiamondPayTransactionStatus];


export const DiamondPayTransactionStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  failed: 'failed',
  expired: 'expired',
} as const;

export interface DiamondPayTransaction {
  transactionId: string;
  status: DiamondPayTransactionStatus;
  amount: number;
  /** @nullable */
  qrCodeUrl?: string | null;
  /** @nullable */
  qrCodeData?: string | null;
  /** @nullable */
  paymentUrl?: string | null;
  createdAt?: string;
}

export type ListProductsParams = {
search?: string;
category?: string;
lowStock?: boolean;
};

export type ListSalesParams = {
dateFrom?: string;
dateTo?: string;
limit?: number;
};

export type ListCustomersParams = {
search?: string;
};

export type ListDebtsParams = {
customerId?: number;
status?: ListDebtsStatus;
};

export type ListDebtsStatus = typeof ListDebtsStatus[keyof typeof ListDebtsStatus];


export const ListDebtsStatus = {
  pending: 'pending',
  partial: 'partial',
  paid: 'paid',
} as const;

export type GetTopProductsParams = {
limit?: number;
};

export type GetSalesByPeriodParams = {
period?: GetSalesByPeriodPeriod;
};

export type GetSalesByPeriodPeriod = typeof GetSalesByPeriodPeriod[keyof typeof GetSalesByPeriodPeriod];


export const GetSalesByPeriodPeriod = {
  day: 'day',
  week: 'week',
  month: 'month',
} as const;