# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LIMAPAY API - Gestion commerciale pour commerces physiques du Sénégal
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Créer un compte
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Compte créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Données invalides
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Se connecter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Connecté
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Identifiants incorrects
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Profil utilisateur courant
      responses:
        "200":
          description: Profil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Non authentifié

  # PRODUCTS
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: Lister les clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Liste des clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Créer un client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Client créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Obtenir un client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Modifier un client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Client modifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Supprimer un client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Supprimé

  # DEBTS
...
  /dashboard/top-products:
    get:
      operationId: getTopProducts
      tags: [dashboard]
      summary: Produits les plus vendus
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top produits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
  /dashboard/sales-by-period:
    get:
      operationId: getSalesByPeriod
      tags: [dashboard]
      summary: Ventes par période
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month]
      responses:
        "200":
          description: Ventes par période
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesPeriodPoint"

  # STOCK
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, shopName]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        shopName:
          type: string
        role:
          type: string
          enum: [owner, cashier]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role, shopName]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [owner, cashier]
        shopName:
          type: string
        createdAt:
          type: string
          format: date-time

    # PRODUCTS
...
    Customer:
      type: object
      required: [id, name, phone]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        totalDebt:
          type: number
        createdAt:
          type: string
          format: date-time
    CustomerInput:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string

    # DEBTS
...
    TopProduct:
      type: object
      required: [productId, productName, totalQuantity, totalRevenue]
      properties:
        productId:
          type: integer
        productName:
          type: string
        totalQuantity:
          type: integer
        totalRevenue:
          type: number
    SalesPeriodPoint:
      type: object
      required: [label, revenue, salesCount]
      properties:
        label:
          type: string
        revenue:
          type: number
        salesCount:
          type: integer

    # STOCK
...